removing CommandsContext
