+++ crabzilla-vertx-pgclient/src/test/java/io/github/crabzilla/pgclient/command/CommandsNotificationIT.kt
@@ -24,7 +24,6 @@ import java.util.UUID
 class CommandsNotificationIT {
 
   private lateinit var jsonSerDer: JsonSerDer
-  private lateinit var commandsContext: CommandsContext
   private lateinit var commandController: CommandController<Customer, CustomerCommand, CustomerEvent>
   private lateinit var repository: SnapshotTestRepository<Customer>
   private lateinit var testRepo: TestRepository
@@ -33,7 +32,6 @@ class CommandsNotificationIT {
   fun setup(vertx: Vertx, tc: VertxTestContext) {
     jsonSerDer = KotlinJsonSerDer(example1Json)
     val pgPool = pgPool(vertx)
-    commandsContext = CommandsContext(vertx, jsonSerDer, pgPool)
     val snapshotRepo2 = PersistentSnapshotRepo<Customer, CustomerEvent>(customerConfig.name, jsonSerDer)
     commandController = CommandController(
       vertx = vertx, pgPool = pgPool, jsonSerDer = jsonSerDer,
