+++ crabzilla-vertx-pgclient/src/test/java/io/github/crabzilla/pgclient/command/CorrelationCausationIdsIT.kt
@@ -27,7 +27,6 @@ import java.util.UUID
 class CorrelationCausationIdsIT {
 
   private lateinit var jsonSerDer: JsonSerDer
-  private lateinit var commandsContext: CommandsContext
   private lateinit var commandController: CommandController<Customer, CustomerCommand, CustomerEvent>
   private lateinit var repository: SnapshotTestRepository<Customer>
   private lateinit var testRepo: TestRepository
@@ -36,7 +35,6 @@ class CorrelationCausationIdsIT {
   fun setup(vertx: Vertx, tc: VertxTestContext) {
     jsonSerDer = KotlinJsonSerDer(example1Json)
     val pgPool = pgPool(vertx)
-    commandsContext = CommandsContext(vertx, jsonSerDer, pgPool)
     val snapshotRepo2 = PersistentSnapshotRepo<Customer, CustomerEvent>(customerConfig.name, jsonSerDer)
     commandController = CommandController(vertx, pgPool, jsonSerDer, customerConfig, snapshotRepo2,)
     repository = SnapshotTestRepository(pgPool, example1Json)
